# network: only make progress when card reports new stuff happened